requirements.txt
pyyaml
pydantic
apscheduler
func-timeout
requests
beautifulsoup4
pypdf[image]
oss2
markdown2
python-telegram-bot
